% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_level_na.R
\name{add_level_na}
\alias{add_level_na}
\title{A add_level_na function}
\usage{
add_level_na(x)
}
\description{
This function allows you to add the level "not available" to a factor variable.
}
\examples{
add_level_na()
}
\keyword{available,}
\keyword{level}
\keyword{na,}
\keyword{not}
