% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_month.R
\name{add_month}
\alias{add_month}
\title{A function to add variables of month and year to a date variable}
\usage{
add_month(data, name)
}
\description{
Add several variables to a dataframe on a date variable with a name on a pattern "date_dob", month, year, yearmonth,weeday
}
\keyword{add_month}
