% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_docx.R
\name{attach_docx}
\alias{attach_docx}
\title{A attach_docx Function}
\usage{
attach_docx(obj, file = "table1.docx")
}
\description{
This function allows you to attach a docx file of a table to the html output.
}
\keyword{a}
\keyword{attach}
\keyword{docx}
\keyword{file}
\keyword{html}
\keyword{output}
\keyword{to}
