% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.R
\name{capwords}
\alias{capwords}
\title{A capwords Function}
\usage{
capwords(s, strict = FALSE)
}
\description{
This function allows you to capitalize your words.
}
\examples{
capwords()
}
\keyword{capitalize}
