% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_to_str.R
\name{ci_to_str}
\alias{ci_to_str}
\title{A ci_to_str Function}
\usage{
ci_to_str(fit, lower, upper, digits = 2)
}
\description{
This function allows you to build a text string of confidence interver.
}
\examples{
ci_to_str()
}
\keyword{ci}
\keyword{confidence}
\keyword{interval,}
