% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cors_mixed.R
\name{cors_mixed}
\alias{cors_mixed}
\title{A correlation between multiple variables}
\usage{
cors_mixed(df)
}
\description{
This function allows you to calculate correlation between multiple variables mixed with continuous and categorical variables.
}
\examples{
cors_mixed()
}
\keyword{correlation}
