% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_single_value_column.R
\name{delete_single_value_column}
\alias{delete_single_value_column}
\title{A delete columns that all rows have only one value}
\usage{
delete_single_value_column(obj)
}
\description{
This function allows you to delete all columns in a data.frame or data.table that have only one unique values (missing value included)
}
\keyword{columns}
\keyword{delete}
\keyword{single}
\keyword{value}
