% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deming_regression.R
\name{deming_regression}
\alias{deming_regression}
\title{Deming Regression Function}
\usage{
deming_regression(x, y, lambda = 1)
}
\description{
This function allows you to run Deming regression.
}
\keyword{Deming}
\keyword{regression}
