% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_cont.R
\name{desc_cont}
\alias{desc_cont}
\title{A desc_cont Function}
\usage{
desc_cont(x, digits_p = 1, digits_c = 3)
}
\description{
Describe a continuous variable.
}
\examples{
x <- c(rnorm(100), rep(NA, 20))
desc_cont(x)
}
\keyword{continuous}
\keyword{describe}
\keyword{variable}
