% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_levels.R
\name{drop_levels}
\alias{drop_levels}
\title{The Drop Levels Functions}
\usage{
drop_levels(df)
}
\description{
This function allows you to drop all unused levels in a data.frame.
}
\examples{

drop_levels()
}
\keyword{drop}
\keyword{levels}
