% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{flowchart}
\alias{flowchart}
\title{A flowchart Function}
\usage{
flowchart(
  data,
  vars,
  top = "Total",
  font_size = 15,
  icon_size = 40,
  edge_width = 2,
  plot_height = "800px",
  plot_width = "100\%"
)
}
\description{
This function allows you to create a flowchart
}
\keyword{flowchart}
