% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_po.R
\name{get_cdf}
\alias{get_cdf}
\title{A plot proportional odds function}
\usage{
get_cdf(data, ordinal, group)
}
\description{
This function allows you to plot propotional odds compare two groups on logit of accumulative proportions
}
\keyword{odds}
\keyword{plot}
\keyword{proportional}
