% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_name.R
\name{get_coef_name}
\alias{get_coef_name}
\title{A get coefficients' names Function}
\usage{
get_coef_name(obj)
}
\description{
This function allows you to get all coefficients names from a dataframe.
}
\keyword{coefficients}
