% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levels.R
\name{get_levels}
\alias{get_levels}
\title{A function to get levels of all variables of a data.frame}
\usage{
get_levels(data, vars = NULL)
}
\description{
This function allows you to get levels of all variables of a data.frame.
}
\keyword{levels}
