% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ors.R
\name{get_or}
\alias{get_or}
\title{A function to get odds ratios on a binary outcome}
\usage{
get_or(y, x, data, digits = 2)
}
\description{
This function allows you to get related measurements on a binary outcome
}
\keyword{analysis}
\keyword{odds}
\keyword{ratio,}
\keyword{univariate}
