% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_hist_prop.R
\name{get_range}
\alias{get_range}
\title{This function allows you to plot a histogram along with a percentage change}
\usage{
get_range(x, digits = 2, pad = 0.5)
}
\description{
This function allows you to plot a histogram along with a percentage change
}
\keyword{histogram}
\keyword{percentage}
