% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{get_tableone}
\alias{get_tableone}
\title{A Table1 Function}
\usage{
get_tableone(data, vars = Vars, factorVars = FactorVars, ...)
}
\description{
This function allows you to create table one.
}
\examples{
Table1()
}
\keyword{one}
\keyword{table}
