% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_varimp.R
\name{get_varimp.glm}
\alias{get_varimp.glm}
\title{A variable importance plot function}
\usage{
get_varimp.glm(mod)
}
\description{
This function to plot variable importance.
}
\keyword{importance,}
\keyword{plot}
\keyword{variable}
