% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{A heatmap Function}
\usage{
heatmap(
  obj,
  title = "Matrix of R-squared",
  width = 800,
  height = 800,
  pallete = "Blues"
)
}
\description{
This function allows you to create a heatmap.
}
\keyword{heatmap}
