% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_ordinal.R
\name{mean_of_ordinal}
\alias{mean_of_ordinal}
\title{A function get mean of ordinals}
\usage{
mean_of_ordinal(yhat, intercepts, ordinals)
}
\description{
This function allows you to get the mean of predicted ordinals from Frank's lrm object.
yhat is the linear predictions on the first intercept (second of the ranks)
}
\keyword{mean}
\keyword{ordinal}
