% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ci.R
\name{model_ci}
\alias{model_ci}
\title{A Function to get coefficients and their confidence intervals from a model object}
\usage{
model_ci(obj, digits = 2, ...)
}
\description{
This function allows you to get coefficients and their confidence intervals from a model object.
}
\keyword{coefficient}
\keyword{confidence}
\keyword{interval,}
