% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_coef.R
\name{model_coef}
\alias{model_coef}
\title{A Function to get the coefficent names from a model object}
\usage{
model_coef(x, ...)
}
\description{
This function allows you to catch the names of coefficients from a model object
}
\examples{
model_coef()
}
\keyword{coefficients}
\keyword{model}
