% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{A Function to get the data.frame from a model object}
\usage{
model_data(x, ...)
}
\description{
This function allows you to catch the data.frame from a model object
}
\keyword{data.frame}
\keyword{model}
