% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vars.R
\name{model_vars}
\alias{model_vars}
\title{A Function to get the variable names from a model object}
\usage{
model_vars(x, ...)
}
\description{
This function allows you to catch the names of variables from a model object
}
\keyword{model}
\keyword{variables}
