% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot}
\alias{one_hot}
\title{A one-hot function}
\usage{
one_hot(data, vars_numeric, vars_factor)
}
\description{
This function allows you to transform variables into dummy variables.
}
\examples{
one_hot()
}
\keyword{dummy}
\keyword{one-hot}
\keyword{variable}
