% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{A percent Function}
\usage{
percent(x, digits = 2, format = "f", ...)
}
\description{
This function allows you to format a number as a percentage.
}
\examples{
percent()
}
\keyword{format}
\keyword{percent}
\keyword{percentage}
