% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{A plot_hist Function}
\usage{
plot_hist(x)
}
\description{
Describe a continuous variable with a histogram.
}
\examples{
x <- c(rnorm(100), rep(NA, 20))
plot_hist(x)
}
\keyword{a}
\keyword{continuous}
\keyword{histogram}
\keyword{of}
\keyword{plot}
\keyword{variable}
