% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nonlinear.R
\name{plot_nonlinear}
\alias{plot_nonlinear}
\title{A plot function for nonlinear relationship}
\usage{
plot_nonlinear(y, x, data, k = 10, family = "binomial")
}
\description{
This function allows you to plot a graph on the nonlinear relationship of two variables.
}
\examples{
plot_nonlinear()
}
\keyword{plot_nonlinear}
