% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nonlinear_b.R
\name{plot_nonlinear_b}
\alias{plot_nonlinear_b}
\title{A plot function for nonlinear relationship}
\usage{
plot_nonlinear_b(y, x, data, k = 10, family = "binomial")
}
\description{
This function allows you to plot a graph on the nonlinear relationship of two variables based on binned intervals of the x variable.
}
\keyword{plot_nonlinear_b}
