% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_calibration.R
\name{plt_calibration}
\alias{plt_calibration}
\title{A Calibration Plot Function}
\usage{
plt_calibration(response, prediction, n)
}
\description{
This function allows you to plot observed risk (rolling average) against predicted risk (rolling average from the lowerest predicted probability to the highest).
}
\keyword{calibration}
\keyword{plot}
