% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtColNames.R
\name{prtColNames}
\alias{prtColNames}
\title{A print column names Function}
\usage{
prtColNames(data, type = NULL)
}
\description{
This function allows you to print column names.
}
\keyword{column}
\keyword{names}
\keyword{print}
