% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrtLabels.R
\name{PrtLabels}
\alias{PrtLabels}
\title{A PrtLabels Function}
\usage{
PrtLabels(data, copy = TRUE, print = FALSE)
}
\description{
This function allows you to print column labels.
}
\examples{
PrtLabels
}
\keyword{column}
\keyword{label}
\keyword{print}
