% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrtTable1.R
\name{PrtTable1}
\alias{PrtTable1}
\title{A PrtTable1 Function}
\usage{
PrtTable1()
}
\description{
This function allows you to print table one.
}
\examples{
PrtTable1
}
\keyword{one}
\keyword{print}
\keyword{table}
