% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_description.R
\name{pvalue_description}
\alias{pvalue_description}
\title{A function describes how p values being calcualted}
\usage{

  pvalue_description(text = c("The default test for categorical variables is chi-squared test with Yates correction for continuity.",
  "The default test for continuous variables is one-way ANOVA on means assuming equal variance. Kruskal-Wallis rank sum test has been applied for the variables where median and interquartile range (IQR) are show on the table."),
  notation = "symbol")
}
\description{
This function describes how p values being calcualted
}
\examples{
describe_pvalue()
}
\keyword{p-value}
