% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameVariables.R
\name{renameVariables}
\alias{renameVariables}
\title{A renameVariables Function}
\usage{
renameVariables(obj, lst)
}
\description{
This function allows you to rename your variable names.
}
\keyword{name}
\keyword{rename}
\keyword{variable}
