% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_regression.R
\name{residual_regression}
\alias{residual_regression}
\title{A residual regression function}
\usage{
residual_regression(data, outcome_x, outcome_y, predictors)
}
\description{
Deming regression on two residuals of linear regression
}
\keyword{regression}
\keyword{residual}
