% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_impute.R
\name{single_impute}
\alias{single_impute}
\title{A single imputeFunction}
\usage{
single_impute(vctr)
}
\description{
This function single impute mean for missing values for continuous variable and add na as a category for missing values in categorical variable.
}
\keyword{impute}
