% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl.R
\name{stl}
\alias{stl}
\title{A style Function}
\usage{
stl(full_width = FALSE, bootstrap_options = c("striped", "hover",
  "condensed", "responsive"), position = "left", fixed_thead = TRUE,
  ...)
}
\description{
This function allows you to style your html output.
}
\examples{
stl()
}
\keyword{html}
\keyword{style}
