% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{styling}
\alias{styling}
\title{A styling Function}
\usage{
styling(
  ...,
  full_width = FALSE,
  bootstrap_options = c("striped", "hover", "condensed", "responsive"),
  position = "left",
  fixed_thead = TRUE
)
}
\description{
This function allows you to styling your html output.
}
\keyword{html}
\keyword{style}
