% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_desc.R
\name{sv_desc}
\alias{sv_desc}
\title{A sv_desc Function}
\usage{
sv_desc(data, file_name = "Description_of_Data_Field.html")
}
\description{
This function allows you to save (output) a html file that describes fields in the input dataset.
}
\keyword{dataset}
\keyword{describe}
\keyword{field}
