% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tab2.R
\name{tab_freq}
\alias{tab_freq}
\title{This function allows you to get a tabulation of two categorical variables.}
\usage{
tab_freq(outcome, group, data, digits = 2, digits.pct = 1,
  label = NA)
}
\description{
This function allows you to get a tabulation of two categorical variables.
}
\keyword{tabulate}
