% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{Table1}
\alias{Table1}
\title{A Table1 Function}
\usage{
Table1(data, Vars, FactorVars, NonNormals = NULL, Strata,
  contDigits = 1, minMax = FALSE, ...)
}
\description{
This function allows you to create table one.
}
\examples{
Table1()
}
\keyword{one}
\keyword{table}
