% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table2_nbyc.R
\name{Table2_nbyc}
\alias{Table2_nbyc}
\title{A correlation function}
\usage{
Table2_nbyc(data, var, groups, digits_p = 1, digits_c = 3)
}
\description{
This function allows you to get the basis discriptive statistics from a continuous variable against a set of categorical variables
}
\keyword{ANOVA,}
\keyword{Kruskal}
\keyword{Wallis}
\keyword{mean,}
