% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table2_nbyn.R
\name{Table2_nbyn}
\alias{Table2_nbyn}
\title{A correlation function}
\usage{
Table2_nbyn(data, var, nvars, digits_p = 3, digits_n = 3, digits_pct = 1)
}
\description{
This function allows you to get the correlation coefficients from a continuous variable against a set of continuous variables
}
\keyword{Cor,}
\keyword{Peason's}
\keyword{Spearman's}
\keyword{correlation,}
\keyword{rho}
