% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Utilities}
\alias{Utilities}
\title{Utility Functions}
\usage{
Utilities()
}
\description{
This function allows you to negate the \%in\% function.
}
\examples{
\\\%notin\\\%


}
\keyword{\%in\%}
\keyword{negate}
