% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtable.R
\name{wtable}
\alias{wtable}
\title{A weighted table() function}
\usage{
wtable(data, rowvar, colvar, weightvar = 1, rowlabel = NULL,
  collabel = NULL)
}
\description{
Tabulate two factors with weight.
}
\keyword{wtable}
