% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_formula.R
\name{wu_formula}
\alias{wu_formula}
\title{A formula Function}
\usage{
wu_formula(outcome, predictors)
}
\description{
This function allows you to create a formula by an outcome and a vector of predictors.
}
\examples{
wu_formula()
}
\keyword{formula}
