% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_month.R
\name{add_month}
\alias{add_month}
\title{A function to add variables of month and year to a date variable}
\usage{
add_month(data, name)
}
\description{
This function allows you to add several variables to the dataframe
}
\keyword{add_month}
