% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mixed.R
\name{cor_mixed}
\alias{cor_mixed}
\title{A correlation between two variables}
\usage{
cor_mixed(var1, var2)
}
\description{
This function allows you to calculate correlation between two variables mixed with continuous and categorical variables.
If there are two continuous variables or one continuous and one categorical varialbe, the function returns r squared of the linear regression.
If there are two categorical variables, the function returns Cramer's V squared.
}
\examples{
cor_mixed()
}
\keyword{correlation}
