% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{flowchart}
\alias{flowchart}
\title{A flowchart Function}
\usage{
flowchart(data, vars, top = "Total", font_size = 30, icon_size = 80,
  edge_width = 10, plot_height = "800px", plot_width = "100\%")
}
\description{
This function allows you to create a flowchart
}
\examples{
flowchart()
}
\keyword{flowchart}
