% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Utility Functions}
\usage{
... \%notin\% NA
}
\description{
This function allows you to negate the %in% function.
}
\examples{
\%notin\%


}
\keyword{%in%}
\keyword{negate}
