% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne.R
\name{\%notin\%}
\alias{\%notin\%}
\title{A %notin% Function}
\usage{
... \%notin\% NA
}
\description{
This function allows you to negate the %in% function.
}
\examples{
\%notin\%



}
\keyword{%in%}
\keyword{negate}
