% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{A heatmap Function}
\usage{
heatmap(obj, title = "", width = 800, height = 800)
}
\description{
This function allows you to create a heatmap.
}
\examples{
heatmap()
}
\keyword{heatmap}
