% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrtColNames.R, R/prtColNames.R
\name{PrtColNames}
\alias{PrtColNames}
\title{A PrtColNames Function}
\usage{
PrtColNames(obj)

PrtColNames(obj)
}
\description{
This function allows you to print column names.

This function allows you to print column names.
}
\examples{
PrtColNames
prtColNames()
}
\keyword{column}
\keyword{name}
\keyword{names}
\keyword{print}
