% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrtColNames.R
\name{PrtColNames}
\alias{PrtColNames}
\title{A PrtColNames Function}
\usage{
PrtColNames(data, type = NULL)
}
\description{
This function allows you to print column names.
}
\examples{
PrtColNames
}
\keyword{column}
\keyword{name}
\keyword{print}
